chore: move check components
