+++ ui/src/alerting/components/builder/AlertBuilder.tsx
@@ -3,10 +3,10 @@ import React, {FC} from 'react'
 
 // Components
 import FancyScrollbar from 'src/shared/components/fancy_scrollbar/FancyScrollbar'
-import CheckMetaCard from 'src/alerting/components/builder/CheckMetaCard'
-import CheckMessageCard from 'src/alerting/components/builder/CheckMessageCard'
-import CheckConditionsCard from 'src/alerting/components/builder/CheckConditionsCard'
-import CheckMatchingRulesCard from 'src/alerting/components/builder/CheckMatchingRulesCard'
+import CheckMetaCard from 'src/checks/components/CheckMetaCard'
+import CheckMessageCard from 'src/checks/components/CheckMessageCard'
+import CheckConditionsCard from 'src/checks/components/CheckConditionsCard'
+import CheckMatchingRulesCard from 'src/checks/components/CheckMatchingRulesCard'
 import {isFlagEnabled} from 'src/shared/utils/featureFlag'
 
 const AlertBuilder: FC = () => {
