+++ ui/src/checks/components/CheckConditionsCard.tsx
