+++ ui/src/checks/components/CheckHistory.scss
