+++ ui/src/checks/components/CheckHistory.tsx
@@ -5,8 +5,8 @@ import {connect} from 'react-redux'
 // Components
 import {Page} from '@influxdata/clockface'
 import EventViewer from 'src/eventViewer/components/EventViewer'
-import CheckHistoryControls from 'src/alerting/components/CheckHistoryControls'
-import CheckHistoryVisualization from 'src/alerting/components/CheckHistoryVisualization'
+import CheckHistoryControls from 'src/checks/components/CheckHistoryControls'
+import CheckHistoryVisualization from 'src/checks/components/CheckHistoryVisualization'
 import AlertHistoryQueryParams from 'src/alerting/components/AlertHistoryQueryParams'
 import EventTable from 'src/eventViewer/components/EventTable'
 import GetResources from 'src/resources/components/GetResources'
