+++ ui/src/checks/components/CheckHistoryVisualization.tsx
