+++ ui/src/checks/components/CheckLevelsDropdown.tsx
