+++ ui/src/checks/components/CheckMetaCard.tsx
@@ -5,7 +5,7 @@ import {connect} from 'react-redux'
 // Components
 import {Form, ComponentSize, ComponentColor, Grid} from '@influxdata/clockface'
 import DashedButton from 'src/shared/components/dashed_button/DashedButton'
-import CheckTagRow from 'src/alerting/components/builder/CheckTagRow'
+import CheckTagRow from 'src/checks/components/CheckTagRow'
 import BuilderCard from 'src/timeMachine/components/builderCard/BuilderCard'
 import DurationInput from 'src/shared/components/DurationInput'
