+++ ui/src/index.tsx
@@ -77,14 +77,14 @@ import RenameVariableOverlay from 'src/variables/components/RenameVariableOverla
 import UpdateVariableOverlay from 'src/variables/components/UpdateVariableOverlay'
 import TaskImportFromTemplateOverlay from 'src/tasks/components/TaskImportFromTemplateOverlay'
 import StaticTemplateViewOverlay from 'src/templates/components/StaticTemplateViewOverlay'
-import CheckHistory from 'src/alerting/components/CheckHistory'
+import CheckHistory from 'src/checks/components/CheckHistory'
 import AlertingIndex from 'src/alerting/components/AlertingIndex'
 import AlertHistoryIndex from 'src/alerting/components/AlertHistoryIndex'
 import BucketsDeleteDataOverlay from 'src/shared/components/DeleteDataOverlay'
 import DEDeleteDataOverlay from 'src/dataExplorer/components/DeleteDataOverlay'
-import NewThresholdCheckEO from 'src/alerting/components/NewThresholdCheckEO'
-import NewDeadmanCheckEO from 'src/alerting/components/NewDeadmanCheckEO'
-import EditCheckEO from 'src/alerting/components/EditCheckEO'
+import NewThresholdCheckEO from 'src/checks/components/NewThresholdCheckEO'
+import NewDeadmanCheckEO from 'src/checks/components/NewDeadmanCheckEO'
+import EditCheckEO from 'src/checks/components/EditCheckEO'
 import NewRuleOverlay from 'src/notifications/rules/components/NewRuleOverlay'
 import EditRuleOverlay from 'src/notifications/rules/components/EditRuleOverlay'
 import NewEndpointOverlay from 'src/notifications/endpoints/components/NewEndpointOverlay'
