+++ ui/src/style/chronograf.scss
@@ -118,7 +118,7 @@
 @import 'src/notifications/rules/components/NewRuleOverlay.scss';
 @import 'src/shared/components/dashed_button/DashedButton.scss';
 @import 'src/alerting/components/AlertingIndex.scss';
-@import 'src/alerting/components/CheckHistory.scss';
+@import 'src/checks/components/CheckHistory.scss';
 @import 'src/alerting/components/builder/AlertBuilder.scss';
 @import 'src/eventViewer/components/EventTable.scss';
 @import 'src/alerting/components/AlertHistoryIndex.scss';
